# Remove unnecessary conversions on instructions