--- a/compiler/bytecode/generator.go
+++ b/compiler/bytecode/generator.go
@@ -1,9 +1,6 @@
 package bytecode
 
 import (
-	"bytes"
-	"strings"
-
 	"github.com/goby-lang/goby/compiler/ast"
 )
 
@@ -42,13 +39,6 @@ func (g *Generator) InitTopLevelScope(program *ast.Program) {
 	g.scope = &scope{program: program, localTable: newLocalTable(0), anchors: make(map[string]*anchor)}
 }
 
-// GenerateByteCode returns compiled instructions in string format
-func (g *Generator) GenerateByteCode(stmts []ast.Statement) string {
-	g.compileStatements(stmts, g.scope, g.scope.localTable)
-
-	return strings.TrimSpace(strings.Replace(g.instructionsToString(), "\n\n", "\n", -1))
-}
-
 // GenerateInstructions returns compiled instructions
 func (g *Generator) GenerateInstructions(stmts []ast.Statement) []*InstructionSet {
 	g.compileStatements(stmts, g.scope, g.scope.localTable)
@@ -56,16 +46,6 @@ func (g *Generator) GenerateInstructions(stmts []ast.Statement) []*InstructionSe
 	//fmt.Println(g.instructionsToString())
 	//fmt.Print()
 	return g.instructionSets
-}
-
-func (g *Generator) instructionsToString() string {
-	var out bytes.Buffer
-
-	for _, is := range g.instructionSets {
-		out.WriteString(is.compile())
-	}
-
-	return out.String()
 }
 
 func (g *Generator) compileCodeBlock(is *InstructionSet, stmt *ast.BlockStatement, scope *scope, table *localTable) {