--- a/compiler/compiler.go
+++ b/compiler/compiler.go
@@ -7,19 +7,6 @@ import (
 	"github.com/goby-lang/goby/compiler/parser"
 )
 
-// CompileToBytecode compiles input source code into Goby bytecode
-func CompileToBytecode(input string) (string, error) {
-	l := lexer.New(input)
-	p := parser.New(l)
-	program, err := p.ParseProgram()
-	if err != nil {
-		return "", fmt.Errorf(err.Message)
-	}
-	g := bytecode.NewGenerator()
-	g.InitTopLevelScope(program)
-	return g.GenerateByteCode(program.Statements), nil
-}
-
 // CompileToInstructions compiles input source code into instruction set data structures
 func CompileToInstructions(input string, parserMode int) ([]*bytecode.InstructionSet, error) {
 	l := lexer.New(input)