… compiler/bytecode/expression_generation_test.go
… compiler/bytecode/generator_test.go
… compiler/bytecode/statement_generation_test.go
… native/ripper/ripper_test.go
… vm/array_enumerator_test.go
… vm/benchmark_test.go
… vm/inspection_methods.go
… vm/instruction.go
… vm/instruction_translator.go
… vm/method.go
… vm/object.go
… vm/thread.go