We used to convert all types of data into string when compiling instructions, and convert them back to their original data type in vm. This was for the goal of compiling to binary, and since we're not going to do that anymore we should skip this conversion.

And as @eliothedeman said before, this does improve the performance (100% improvement!):

```
benchmarking 0772826a0a295544b55e35b10f10877af452355e
benchmarking a200a8af3b0cf0a7f4546291a17e45c4a8b99078
x86_64-darwin17
benchmark                         old ns/op     new ns/op     delta
BenchmarkBasicMath/add-4          5139          2810          -45.32%
BenchmarkBasicMath/subtract-4     5206          2798          -46.25%
BenchmarkBasicMath/multiply-4     5158          2813          -45.46%
BenchmarkBasicMath/divide-4       5184          2807          -45.85%

benchmark                         old allocs     new allocs     delta
BenchmarkBasicMath/add-4          75             34             -54.67%
BenchmarkBasicMath/subtract-4     75             34             -54.67%
BenchmarkBasicMath/multiply-4     75             34             -54.67%
BenchmarkBasicMath/divide-4       75             34             -54.67%

benchmark                         old bytes     new bytes     delta
BenchmarkBasicMath/add-4          3152          1576          -50.00%
BenchmarkBasicMath/subtract-4     3152          1576          -50.00%
BenchmarkBasicMath/multiply-4     3152          1576          -50.00%
BenchmarkBasicMath/divide-4       3152          1576          -50.00%
```